requirements.txt
pyTelegramBotAPI==3.7.3
Flask==1.1.2
psycopg2==2.8.5
python-telegram-bot-calendar==1.0.2
requests==2.24.0